# artifacts/api-server/src/routes/reports.ts
import { Router } from "express";
import { db } from "@workspace/db";
import {
  auditReportsTable,
  structuralAnomaliesTable,
  ppeViolationsTable,
  workersTable,
} from "@workspace/db";
import { eq, sql } from "drizzle-orm";

const router = Router();

router.get("/reports", async (req, res) => {
  try {
    const reports = await db
      .select()
      .from(auditReportsTable)
      .orderBy(sql`${auditReportsTable.generatedAt} desc`);

    res.json(
      reports.map((r) => ({
        ...r,
        recommendations: r.recommendations as string[],
        generatedAt: r.generatedAt?.toISOString(),
      }))
    );
  } catch (err) {
    req.log.error({ err }, "Error listing reports");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/reports/:reportId", async (req, res) => {
  try {
    const reportId = parseInt(req.params.reportId);
    const [report] = await db
      .select()
      .from(auditReportsTable)
      .where(eq(auditReportsTable.id, reportId));

    if (!report) return res.status(404).json({ error: "Report not found" });

    res.json({
      ...report,
      recommendations: report.recommendations as string[],
      generatedAt: report.generatedAt?.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Error getting report");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/reports", async (req, res) => {
  try {
    const {
      period,
      includeModuleA = true,
      includeModuleB = true,
      includeModuleC = true,
    } = req.body;

    const [anomalies, violations, workers] = await Promise.all([
      db.select().from(structuralAnomaliesTable).where(eq(structuralAnomaliesTable.resolved, false)),
      db.select().from(ppeViolationsTable).where(eq(ppeViolationsTable.resolved, false)),
      db.select().from(workersTable),
    ]);

    const criticalAnomalies = anomalies.filter((a) => a.severity === "critical").length;
    const highAnomalies = anomalies.filter((a) => a.severity === "high").length;
    const totalWorkers = workers.length || 1;
    const violatingWorkers = workers.filter((w) => w.ppeStatus === "violation").length;
    const ppeComplianceRate = ((totalWorkers - violatingWorkers) / totalWorkers) * 100;
    const avgEfficiency = workers.reduce((s, w) => s + w.efficiencyScore, 0) / totalWorkers;

    const costImpact = anomalies.reduce((s, a) => {
      const m =
        a.severity === "critical" ? 15000 :
        a.severity === "high" ? 8000 :
        a.severity === "medium" ? 3000 : 1000;
      return s + m;
    }, 0) + violations.length * 2500;

    const riskLevel =
      criticalAnomalies > 2 || violations.length > 5
        ? "critical"
        : criticalAnomalies > 0 || violations.length > 2
          ? "high"
          : anomalies.length > 3
            ? "medium"
            : "low";

    const structuralSummary = includeModuleA
      ? `Module A scan detected ${anomalies.length} structural anomalies (${criticalAnomalies} critical, ${highAnomalies} high severity). Key deviations include column misalignments and wall offsets. Immediate remediation required for critical elements in Zone-1 and Zone-4. Construction progress at ${48.5 + Math.random() * 5 | 0}% overall with Floor 2 slab showing +${(3 + Math.random() * 2).toFixed(1)}% daily advance.`
      : "Module A scan not included in this report.";

    const safetySummary = includeModuleB
      ? `Module B 360° Guardian detected ${violations.length} active PPE violations across ${[...new Set(violations.map((v) => v.cameraId))].length} camera feeds. PPE compliance rate: ${ppeComplianceRate.toFixed(1)}%. Primary issues: missing helmets (${violations.filter((v) => (v.missingItems as string[]).includes("helmet")).length} workers) and hi-vis vests (${violations.filter((v) => (v.missingItems as string[]).includes("vest")).length} workers). ${criticalAnomalies > 0 ? "CRITICAL: Active danger zone breach detected." : "No active zone breaches at report time."}`
      : "Module B safety scan not included in this report.";

    const efficiencySummary = includeModuleC
      ? `Module C Activity Analyst tracked ${workers.length} workers on site. Team efficiency score: ${avgEfficiency.toFixed(1)}%. ${workers.filter((w) => w.status === "idle").length} workers flagged as idle (>300s without significant movement). Top performers averaging ${(avgEfficiency + 15).toFixed(1)}% efficiency. Estimated labor utilization: ${(avgEfficiency * 0.9).toFixed(1)}%. Recommend targeted check-ins for underperforming workers.`
      : "Module C efficiency scan not included in this report.";

    const fullReport = `# AECI Audit Report — ${period}
Generated: ${new Date().toLocaleString()}

## Executive Summary
This automated audit was generated by the Astra-Eye Construction Intelligence system covering ${period}. The site shows ${riskLevel.toUpperCase()} overall risk level based on structural, safety, and labor analytics.

**Estimated Financial Impact: $${costImpact.toLocaleString()}**

## Module A — Structural Analysis
${structuralSummary}

### Anomaly Details
${anomalies.map((a) => `- [${a.severity.toUpperCase()}] ${a.elementId}: ${a.deviationDescription} (${a.deviationPct}% deviation) in ${a.zone}`).join("\n") || "No anomalies detected."}

## Module B — Safety & PPE Compliance
${safetySummary}

### Violation Details
${violations.slice(0, 5).map((v) => `- Worker in Camera ${v.cameraName}: Missing ${(v.missingItems as string[]).join(", ")} — ${v.severity} severity`).join("\n") || "No violations detected."}

## Module C — Labor Efficiency
${efficiencySummary}

### Worker Status Breakdown
- Active: ${workers.filter((w) => w.status === "active").length} workers
- Idle: ${workers.filter((w) => w.status === "idle").length} workers  
- Break: ${workers.filter((w) => w.status === "break").length} workers

## Risk Assessment
**Overall Risk Level: ${riskLevel.toUpperCase()}**

Financial impact projection based on current deviations and safety violations. Every 1% structural deviation caught early reduces rework costs by an estimated $8,000–$15,000 per element. PPE violations carry liability exposure of $2,500+ per incident under standard construction safety regulations.

## Recommendations
See below for prioritized action items.`;

    const recommendations = [
      criticalAnomalies > 0
        ? `URGENT: Halt work in Zone-1 until critical beam deviation is assessed by structural engineer`
        : "Continue scheduled BIM verification scans daily",
      violations.length > 0
        ? `Conduct mandatory PPE refresher for ${violatingWorkers} non-compliant workers immediately`
        : "Maintain current PPE compliance protocols",
      workers.filter((w) => w.status === "idle").length > 2
        ? `Reassign ${workers.filter((w) => w.status === "idle").length} idle workers to active zones to improve throughput`
        : "Labor efficiency within acceptable parameters",
      "Schedule end-of-week full-site BIM comparison scan",
      "Review and update danger zone perimeters for crane operational area",
    ].filter(Boolean);

    const [report] = await db
      .insert(auditReportsTable)
      .values({
        title: `AECI Audit Report — ${period}`,
        period,
        structuralSummary,
        safetySummary,
        efficiencySummary,
        costImpactEstimate: costImpact,
        riskLevel,
        recommendations,
        fullReport,
      })
      .returning();

    res.status(201).json({
      ...report,
      recommendations: report.recommendations as string[],
      generatedAt: report.generatedAt?.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Error generating report");
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;