# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Astra-Eye Construction Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /module-c/workers/{workerId}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get worker detail with activity log
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated audit reports
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditReport"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a new AI audit report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReportBody"
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditReport"
  /reports/{reportId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific audit report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditReport"
components:
  schemas:
...
    Worker:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        zone:
          type: string
        status:
          type: string
          enum: [active, idle, break, offline]
        efficiencyScore:
          type: number
        movementScore:
          type: number
        idleMinutes:
          type: integer
        ppeStatus:
          type: string
          enum: [compliant, violation, unknown]
        currentCamera:
          type: string
        joinedAt:
          type: string
        activityLog:
          type: array
          items:
            $ref: "#/components/schemas/WorkerActivity"
      required:
        - id
        - name
        - role
        - zone
        - status
        - efficiencyScore
        - movementScore
        - idleMinutes
        - ppeStatus
        - currentCamera
        - joinedAt
    WorkerActivity:
      type: object
      properties:
        time:
          type: string
        action:
          type: string
        duration:
          type: integer
        movementScore:
          type: number
      required: [time, action, duration, movementScore]
...
    AuditReport:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        generatedAt:
          type: string
        period:
          type: string
        structuralSummary:
          type: string
        safetySummary:
          type: string
        efficiencySummary:
          type: string
        costImpactEstimate:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        recommendations:
          type: array
          items:
            type: string
        fullReport:
          type: string
      required:
        - id
        - title
        - generatedAt
        - period
        - structuralSummary
        - safetySummary
        - efficiencySummary
        - costImpactEstimate
        - riskLevel
        - recommendations
        - fullReport
    GenerateReportBody:
      type: object
      properties:
        period:
          type: string
        includeModuleA:
          type: boolean
        includeModuleB:
          type: boolean
        includeModuleC:
          type: boolean
      required: [period]